Add pyproject.toml with PEP-517 build system
python-libpcap depends on Cython to build, but doesn't specify this
build dependency, just documents it in README.rst. This may work if you
install the package into an environment using pip directly, and install
Cython first, but it does not work when using tools which perform builds
in isolated environments, such as PDM, or even just using pip to install
if you don't first install Cython.

Add a pyproject.toml with PEP-517 build backend and build requires to
properly be able to build this using tools which install build
dependencies automatically.